+++ microservices/task/src/main/java/com/solurion/eclipto/task/dto/TaskInfoDto.java
@@ -25,7 +25,7 @@ import jakarta.annotation.Generated;
 @Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.4.0")
 public class TaskInfoDto {
 
-  private Integer id;
+  private Long id;
 
   private String title;
 
@@ -75,9 +75,9 @@ public class TaskInfoDto {
   @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
   private OffsetDateTime dueDate;
 
-  private Integer assignedUserId;
+  private Long assignedUserId;
 
-  private Integer reporterUserId;
+  private Long reporterUserId;
 
   @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
   private OffsetDateTime createdAt;
@@ -92,7 +92,7 @@ public class TaskInfoDto {
   /**
    * Constructor with only required parameters
    */
-  public TaskInfoDto(Integer id, String title, TaskStatusDto status, PriorityEnum priority, Integer reporterUserId, OffsetDateTime createdAt, OffsetDateTime updatedAt) {
+  public TaskInfoDto(Long id, String title, TaskStatusDto status, PriorityEnum priority, Long reporterUserId, OffsetDateTime createdAt, OffsetDateTime updatedAt) {
     this.id = id;
     this.title = title;
     this.status = status;
@@ -102,7 +102,7 @@ public class TaskInfoDto {
     this.updatedAt = updatedAt;
   }
 
-  public TaskInfoDto id(Integer id) {
+  public TaskInfoDto id(Long id) {
     this.id = id;
     return this;
   }
@@ -114,11 +114,11 @@ public class TaskInfoDto {
   @NotNull 
   @Schema(name = "id", example = "7438546582", description = "ID of the task", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("id")
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
@@ -222,7 +222,7 @@ public class TaskInfoDto {
     this.dueDate = dueDate;
   }
 
-  public TaskInfoDto assignedUserId(Integer assignedUserId) {
+  public TaskInfoDto assignedUserId(Long assignedUserId) {
     this.assignedUserId = assignedUserId;
     return this;
   }
@@ -234,15 +234,15 @@ public class TaskInfoDto {
   
   @Schema(name = "assigned_user_id", example = "23486357348", description = "ID of the user who was assigned to the task", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
   @JsonProperty("assigned_user_id")
-  public Integer getAssignedUserId() {
+  public Long getAssignedUserId() {
     return assignedUserId;
   }
 
-  public void setAssignedUserId(Integer assignedUserId) {
+  public void setAssignedUserId(Long assignedUserId) {
     this.assignedUserId = assignedUserId;
   }
 
-  public TaskInfoDto reporterUserId(Integer reporterUserId) {
+  public TaskInfoDto reporterUserId(Long reporterUserId) {
     this.reporterUserId = reporterUserId;
     return this;
   }
@@ -254,11 +254,11 @@ public class TaskInfoDto {
   @NotNull 
   @Schema(name = "reporter_user_id", example = "23486357348", description = "ID of the user who reported the task", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("reporter_user_id")
-  public Integer getReporterUserId() {
+  public Long getReporterUserId() {
     return reporterUserId;
   }
 
-  public void setReporterUserId(Integer reporterUserId) {
+  public void setReporterUserId(Long reporterUserId) {
     this.reporterUserId = reporterUserId;
   }
