+++ microservices/task/src/main/java/com/solurion/eclipto/task/repository/TaskStatusRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskStatusRepository extends JpaRepository<TaskStatusEntity, Integer> {
+public interface TaskStatusRepository extends JpaRepository<TaskStatusEntity, Long> {
 
 }
